Fully finish implementation, but at a price.
Scrambling data is amazing, but the thing is, to actually extract it
back from the images, you need a pretty good system to unravel it.

So far, this implementation scrambles the data up into almost exactly
even pieces. This means that the max data you can store is a function of
the min(all_file_capacities) * image_number, which isn't great. Future
implemenation will have two ways to hide data: In pieces to take up as
much space as possible, or scramble at the cost of space